add `reduce_type()` with all previous existing logic
